Fixed.  AlertDialog Record

+++ app/src/main/java/com/example/link/data/entity/PetEntity.kt
@@ -9,5 +9,5 @@ data class PetEntity (
     val mail : Boolean? = null,
     val type : String? = null,
     val birthDay : List<Int>? = null,
-    val weight : Int? = null
+    val weight : Float? = null
 )
