+++ app/src/main/java/com/example/link/data/repository/PetRepository.kt
@@ -4,7 +4,7 @@ import com.example.link.model.PetModel
 
 interface PetRepository {
 
-    suspend fun savePetData(name: String, isMail: Boolean, type: String, birthday: List<Int>, wight: Int)
+    suspend fun savePetData(name: String, isMail: Boolean, type: String, birthday: List<Int>, wight: Float)
 
     suspend fun  getPetData( userId : String) : PetModel?
